+++ src/main/java/io/nagurea/smsupsdk/campaigns/get/history/ListId.java
@@ -1,0 +1,14 @@
+package io.nagurea.smsupsdk.campaigns.get.history;
+
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@EqualsAndHashCode
+@Builder
+@ToString
+public class ListId {
+    String id;
+}