+++ src/main/java/io/nagurea/smsupsdk/helper/json/typeadapter/SMSUP_FORMATS.java
@@ -1,0 +1,11 @@
+package io.nagurea.smsupsdk.helper.json.typeadapter;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class SMSUP_FORMATS {
+
+    public static final String SMSUP_DATE_FORMAT = "yyyy-MM-dd HH:mm:ss";
+
+}